+++ ui-admin/src/util/tableUtils.tsx
@@ -424,7 +424,7 @@ export type BasicTableConfig<T> = {
 }
 
 /** Default configuration if no `BasicTableConfig` is provided or any of its attributes are not specified. */
-const defaultBasicTableConfig = {
+export const defaultBasicTableConfig = {
   filterable: false
 }
 
@@ -461,100 +461,6 @@ export function basicTableLayout<T>(table: Table<T>, config: BasicTableConfig<T>
       })}
     </tbody>
   </table>
-}
-
-
-// Cell Component
-export const RowDragHandleCell = ({ rowId }: { rowId: string }) => {
-  const { attributes, listeners, isDragging } = useSortable({
-    id: rowId
-  })
-  const style: CSSProperties = {
-    cursor: isDragging ? 'grabbing' : 'grab'
-  }
-
-  return (
-    // Alternatively, you could set these attributes on the rows themselves
-    <button className="ms-2 p-0 btn btn-secondary" style={style} {...attributes} {...listeners}>
-      <FontAwesomeIcon icon={faGripVertical}/>
-    </button>
-  )
-}
-
-// Row Component
-const DraggableRow = function<T>({ row, idFunc }: { row: Row<T>, idFunc: (row: Row<T>) => string | number }) {
-  const { transform, transition, setNodeRef, isDragging } = useSortable({
-    id: idFunc(row)
-  })
-
-  const style: CSSProperties = {
-    transform: CSS.Transform.toString(transform), //let dnd-kit do its thing
-    transition,
-    opacity: isDragging ? 0.8 : 1,
-    zIndex: isDragging ? 1 : 0,
-    position: 'relative'
-  }
-  return (
-    // connect row ref to dnd-kit, apply important styles
-    <tr ref={setNodeRef} style={style}>
-      {row.getVisibleCells().map(cell => (
-        <td key={cell.id} style={{ width: cell.column.getSize(), verticalAlign: 'middle' }}>
-          {flexRender(cell.column.columnDef.cell, cell.getContext())}
-        </td>
-      ))}
-    </tr>
-  )
-}
-
-
-/** helper function for simple table layouts */
-export function useDraggableTableLayout<T>(table: Table<T>, config: BasicTableConfig<T> = {},
-  idList: UniqueIdentifier[], setIdList: Dispatch<SetStateAction<UniqueIdentifier[]>>,
-  idFunc: (row: Row<T>) => UniqueIdentifier) {
-  const { filterable } = { ...defaultBasicTableConfig, ...config }
-  const sensors = useSensors(
-    useSensor(MouseSensor, {}),
-    useSensor(TouchSensor, {}),
-    useSensor(KeyboardSensor, {})
-  )
-
-  // reorder rows after drag & drop
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event
-    if (active && over && active.id !== over.id) {
-      setIdList(data => {
-        const oldIndex = idList.indexOf(active.id)
-        const newIndex = idList.indexOf(over.id)
-        return arrayMove(data, oldIndex, newIndex) //this is just a splice util
-      })
-    }
-  }
-
-  return <DndContext onDragEnd={handleDragEnd} collisionDetection={closestCenter}
-    modifiers={[restrictToVerticalAxis]}
-    sensors={sensors}>
-    <table className={config.tableClass ? config.tableClass : 'table'}>
-      <thead>
-        <tr>
-          {table
-            .getFlatHeaders()
-            .map(header => tableHeader(header, {
-              sortable: true, filterable, useSize: config?.useSize || false
-            }))}
-        </tr>
-      </thead>
-      <tbody>
-        <SortableContext
-          items={idList}
-          strategy={verticalListSortingStrategy}
-        >
-          {table.getRowModel().rows.map(row => (
-            <DraggableRow key={row.id} row={row} idFunc={idFunc} />
-          ))}
-        </SortableContext>
-      </tbody>
-    </table>
-  </DndContext>
 }
 
 /** renders a boolean value as a checkmark (true)  or a blank (false) */
